requirements.txt
allure-pytest==2.5.2
allure-python-commons==2.5.2
apipkg==1.5
atomicwrites==1.2.1
attrs==18.1.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
crayons==0.1.2
enum34==1.1.6
execnet==1.5.0
idna==2.7
lxml==4.2.5
more-itertools==4.3.0
namedlist==1.7
pluggy==0.7.1
py==1.6.0
pytest==3.7.4
pytest-forked==0.2
pytest-rerunfailures==4.1
pytest-xdist==1.23.0
requests==2.19.1
selenium==3.14.0
six==1.11.0
urllib3==1.23
webdriver-manager==1.7
